Removed test requiring linux containers
