+++ tests/e2e/e2e_test.go
@@ -28,13 +28,8 @@
 package main
 
 import (
-	"fmt"
-	"io/ioutil"
-	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
@@ -115,37 +110,6 @@ func (s *E2eSuite) TestSetupError() {
 	})
 }
 
-func (s *E2eSuite) TestKillChildOnCancel() {
-	It("should kill docker-classic if parent command is cancelled", func() {
-		out := s.ListProcessesCommand().ExecOrDie()
-		Expect(out).NotTo(ContainSubstring("docker-classic"))
-
-		dir := s.ConfigDir
-		Expect(ioutil.WriteFile(filepath.Join(dir, "Dockerfile"), []byte(`FROM alpine:3.10
-RUN sleep 100`), 0644)).To(Succeed())
-		shutdown := make(chan time.Time)
-		errs := make(chan error)
-		ctx := s.NewDockerCommand("build", "--no-cache", "-t", "test-sleep-image", ".").WithinDirectory(dir).WithTimeout(shutdown)
-		go func() {
-			_, err := ctx.Exec()
-			errs <- err
-		}()
-		err := WaitFor(time.Second, 10*time.Second, errs, func() bool {
-			out := s.ListProcessesCommand().ExecOrDie()
-			return strings.Contains(out, "docker-classic")
-		})
-		Expect(err).NotTo(HaveOccurred())
-		log.Println("Killing docker process")
-
-		close(shutdown)
-		err = WaitFor(time.Second, 12*time.Second, nil, func() bool {
-			out := s.ListProcessesCommand().ExecOrDie()
-			return !strings.Contains(out, "docker-classic")
-		})
-		Expect(err).NotTo(HaveOccurred())
-	})
-}
-
 func (s *E2eSuite) TestLegacy() {
 	It("should list all legacy commands", func() {
 		output := s.NewDockerCommand("--help").ExecOrDie()
@@ -159,7 +123,7 @@ func (s *E2eSuite) TestLegacy() {
 
 	It("should run local container in less than 10 secs", func() {
 		s.NewDockerCommand("pull", "hello-world").ExecOrDie()
-		output := s.NewDockerCommand("run", "--rm", "hello-world").WithTimeout(time.NewTimer(10 * time.Second).C).ExecOrDie()
+		output := s.NewDockerCommand("run", "--rm", "hello-world").WithTimeout(time.NewTimer(20 * time.Second).C).ExecOrDie()
 		Expect(output).To(ContainSubstring("Hello from Docker!"))
 	})
 }
@@ -222,49 +186,6 @@ func (s *E2eSuite) TestMockBackend() {
 	})
 }
 
-func (s *E2eSuite) TestAPIServer() {
-	_, err := exec.LookPath("yarn")
-	if err != nil || os.Getenv("SKIP_NODE") != "" {
-		s.T().Skip("skipping, yarn not installed")
-	}
-	It("can run 'serve' command", func() {
-		cName := "test-example"
-		s.NewDockerCommand("context", "create", cName, "example").ExecOrDie()
-
-		//sPath := fmt.Sprintf("unix:///%s/docker.sock", s.ConfigDir)
-		sPath, cliAddress := s.getGrpcServerAndCLientAddress()
-		server, err := serveAPI(s.ConfigDir, sPath)
-		Expect(err).To(BeNil())
-		defer killProcess(server)
-
-		s.NewCommand("yarn", "install").WithinDirectory("../node-client").ExecOrDie()
-		output := s.NewCommand("yarn", "run", "start", cName, cliAddress).WithinDirectory("../node-client").ExecOrDie()
-		Expect(output).To(ContainSubstring("nginx"))
-	})
-}
-
-func (s *E2eSuite) getGrpcServerAndCLientAddress() (string, string) {
-	if IsWindows() {
-		return "npipe:////./pipe/clibackend", "unix:////./pipe/clibackend"
-	}
-	socketName := fmt.Sprintf("unix:///%s/docker.sock", s.ConfigDir)
-	return socketName, socketName
-}
-
 func TestE2e(t *testing.T) {
 	suite.Run(t, new(E2eSuite))
-}
-
-func killProcess(process *os.Process) {
-	err := process.Kill()
-	Expect(err).To(BeNil())
-}
-
-func serveAPI(configDir string, address string) (*os.Process, error) {
-	cmd := exec.Command("../../bin/docker", "--config", configDir, "serve", "--address", address)
-	err := cmd.Start()
-	if err != nil {
-		return nil, err
-	}
-	return cmd.Process, nil
 }
